Extract shared mesh LLM types crate
