+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "crates/mesh-llm",
+    "crates/mesh-llm-types",
     "crates/mesh-llm-ui",
     "crates/mesh-llm-plugin",
     "crates/mesh-client",
