+++ README.md
@@ -639,6 +639,7 @@ You can also try the hosted demo:
 - [docs/plugins/README.md](docs/plugins/README.md) for the plugin system and blackboard internals
 - [docs/design/VIRTUAL_LLM.md](docs/design/VIRTUAL_LLM.md) for inter-model collaboration design
 - [crates/mesh-llm/README.md](crates/mesh-llm/README.md) for Rust crate structure
+- [crates/mesh-llm-types/README.md](crates/mesh-llm-types/README.md) for shared protocol-facing model and mesh data types
 - [crates/mesh-llm-ui/README.md](crates/mesh-llm-ui/README.md) for web console source, build, and embedded asset ownership
 - [ROADMAP.md](ROADMAP.md) for future work
