+++ crates/mesh-client/README.md
@@ -19,6 +19,11 @@ Language bindings should generally reach this crate through:
 Keep this crate implementation-focused. Public, app-facing ergonomics should be
 added in `crates/mesh-api/`, not here.
 
+Shared protocol-facing model/type definitions are owned by
+`crates/mesh-llm-types/` and re-exported here where existing client call sites
+expect them. Keep pure shared data there instead of adding host-runtime
+dependencies to this crate.
+
 Client requests should preserve the full model ref chosen by the caller. Model
 resolution, stage topology, and runtime lifecycle remain server-side mesh
 responsibilities.
