+++ crates/mesh-llm-types/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "mesh-llm-types"
+edition.workspace = true
+license.workspace = true
+version.workspace = true
+description = "Shared protocol-facing data types for Mesh LLM crates"
+repository = "https://github.com/Mesh-LLM/mesh-llm"
+homepage = "https://github.com/Mesh-LLM/mesh-llm"
+readme = "README.md"
+
+[dependencies]
+hex = "0.4"
+serde.workspace = true
+serde_json.workspace = true
+sha2.workspace = true
