+++ crates/mesh-llm-types/src/lib.rs
@@ -1,0 +1,4 @@
+#![forbid(unsafe_code)]
+
+pub mod mesh;
+pub mod models;
