+++ crates/mesh-llm-types/src/models/mod.rs
@@ -1,0 +1,8 @@
+pub mod capabilities;
+pub mod topology;
+
+pub use capabilities::{
+    merge_config_signals, merge_name_signals, merge_sibling_signals, CapabilityLevel,
+    ModelCapabilities,
+};
+pub use topology::{ModelMoeInfo, ModelTopology};
