+++ crates/mesh-llm/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 ansi-to-tui = "8"
 mesh-llm-plugin = { path = "../mesh-llm-plugin" }
+mesh-llm-types = { path = "../mesh-llm-types" }
 mesh-llm-ui = { path = "../mesh-llm-ui" }
 mesh-client = { package = "mesh-llm-client", path = "../mesh-client", features = ["host-io"] }
 model-artifact = { path = "../model-artifact" }
