+++ crates/mesh-llm/src/models/topology.rs
@@ -1,6 +1,6 @@
 use std::path::Path;
 
-pub use mesh_client::models::topology::{ModelMoeInfo, ModelTopology};
+pub use mesh_llm_types::models::topology::{ModelMoeInfo, ModelTopology};
 
 #[allow(dead_code)]
 pub fn infer_local_model_topology(_path: &Path) -> Option<ModelTopology> {
